Modified onlineGUi ped_tracking.C and ped_tracking_trigger.C
Set the Y-axis of the pedestal difference plot to the minimum of 15
or 1.2 times the largest difference.

In ped_tracking
---------------
If both golden and current run do not have a pedestal for that channel,
then ped diff is zero.

If only the golden run does not have a pedestal for that channel , then
sets the ped difference to  negative pedestal value for the current run.

If only the current run does not have a pedestal for that channel , then
sets the ped difference to pedestal value for the golden run.